:truck: api: move CompetenceMark from lib to src
