+++ api/tests/integration/infrastructure/repositories/competence-mark-repository_test.js
@@ -1,7 +1,7 @@
 import _ from 'lodash';
 
-import { CompetenceMark } from '../../../../lib/domain/models/CompetenceMark.js';
 import * as competenceMarkRepository from '../../../../lib/infrastructure/repositories/competence-mark-repository.js';
+import { CompetenceMark } from '../../../../src/certification/results/domain/models/CompetenceMark.js';
 import { databaseBuilder, domainBuilder, expect, knex } from '../../../test-helper.js';
 
 describe('Integration | Repository | CompetenceMark', function () {
