+++ api/tests/tooling/domain-builder/factory/build-competence-mark.js
@@ -1,4 +1,4 @@
-import { CompetenceMark } from '../../../../lib/domain/models/CompetenceMark.js';
+import { CompetenceMark } from '../../../../src/certification/results/domain/models/CompetenceMark.js';
 
 const buildCompetenceMark = function ({
   id,
